+++ CHANGELOG.md
@@ -15,6 +15,7 @@ simpler to gain understanding of what you're using.
 * Enhance stacktrace to definition navigation to work for interactively defined vars.
 * New vars: `cider-to-nrepl-filename-function` and `cider-from-nrepl-filename-function`
 are used to translate filenames from/to the nREPL server (default Cygwin implementation provided).
+* Java classpath browser (`M-x cider-classpath`).
 
 ### Changes